Move controller & routing into new common subpackage
Rename ModelServerController to ModelController and move it
together with ModelServerRouting into a new comon subpackage
